[Breaking Change] Firestore array operations (#1269)
This change adds support for adding array elements and removing array elements. It also adds support for querying using the `array-contains` operator.

The bulk of the work is in a refactor of WriteBatch to improve handling of values from the FieldValue statics. The work is based heavily off of the latest iteration of [nodejs-firestore](https://github.com/googleapis/nodejs-firestore/blob/master/dev/src/field-value.ts) and should allow much easier implementation of future DocumentTransform values.

The breaking changes are minor and should not impact users in normal (expected) usage, though should still be noted.

* `Google\Cloud\Firestore\FieldValue` method return values have changed.
* `Google\Cloud\Firestore\WriteBatch::REQUEST_TIME` moved to `Google\Cloud\Firestore\FieldValue\ServerTimestampValue::REQUEST_TIME`.
* `Google\Cloud\Firestore\FieldValue` methods `sentinelValues()` and `isSentinelValue()` have been removed.

cc @schmidt-sebastian